Photo changes and External Ref change
Changed some photos to look more proportionate as well as the colour of the midi box white. I also changed the Cart Controller so that the external ref of an order has a random 8 character string after it, so they should all be unique. I also changed the product description so that in the personaliser it doesn't show the html tags too.