Sorted issue of mapped sku's in personaliser
Before this change once the personaliser was loaded, it had to try find the product in the database to add to the cart rather than it adding the .ALL product. This if statement checks if it's got any info in 'extra' and if it does, grab it, if not it does it the old way.